requirements.txt
Pillow==4.0.0
Requests==2.12.4
